Let's break everything!
The client was previously written with the assumption that the server only
understood two kinds of requests, read and write. Therefore the client was
built layered, with the store layer handling the read/write requests, and their
associated headers, while the layer above that handled everything else. Well
now the server knows about multiappends so these layers don't work. Since the
next series of commits will try to fix that, this is a checkpoint commit so
after I break everything I can roll back.